+++ src/main/java/org/swyp/weddy/domain/storage/web/StorageController.java
@@ -3,14 +3,14 @@ package org.swyp.weddy.domain.storage.web;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-import org.swyp.weddy.domain.storage.service.StorageService;
+import org.swyp.weddy.domain.storage.service.StorageServiceImpl;
 
 @RestController
 @RequestMapping("/api/files")
 public class StorageController implements StorageApiSpec {
-    private final StorageService storageService;
+    private final StorageServiceImpl storageService;
 
-    public StorageController(StorageService storageService) {
+    public StorageController(StorageServiceImpl storageService) {
         this.storageService = storageService;
     }
